feat/#77: 서비스 로직과 리포지토리 분리
