+++ src/idea/repository/idea.repository.ts
@@ -1,5 +1,6 @@
 import { EntityManager } from 'typeorm';
 import { IdeaModel } from '../domain/idea.model';
+import { IdeaOverviewDto } from '../application/dto/response/read-idea-overview.response.dto';
 
 export interface IdeaRepository {
   findAll(manager? : EntityManager): Promise<IdeaModel[]>;
@@ -8,5 +9,14 @@ export interface IdeaRepository {
   findByUserIdAndIsBookmarked(userId: number, manager? : EntityManager): Promise<IdeaModel[]>;
   save(idea: IdeaModel, manager? : EntityManager): Promise<void>;
   delete(id: number, manager? : EntityManager): Promise<void>;
-  getIsBookmarkedByUserIdAndIdeaIds(userId: number, ideaIds: number[], manager? : EntityManager): Promise<{ ideaId: number, isBookmarked: boolean }[]>;
+  findIdeaOverview(
+    page: number,
+    size: number,
+    generation: number,
+    subjectId: number | undefined,
+    isActive: boolean | undefined,
+    isBookmarked: boolean | undefined,
+    userId: number,
+    manager?: EntityManager
+  ): Promise<{ ideas: IdeaOverviewDto[]; totalItems: number }>;
 }
